refactor(rsvp): remove RSVP + nudge, replace with InvitationCard + recipients
RSVP turned out to be mostly overhead — if someone's in, they sign up;
if they're out, they ignore the email. The collect-and-track layer added
complexity without a proportional payoff. Ripping it out cleanly so the
space is open for a better signal later.

Removed
- RSVP tRPC procedures: setRsvpStatus, setRsvpStatusForMember, legacy
  updateRsvp. Dropped rsvp_status from the tripMembers.list SELECT.
- Nudge procedures and email: nudgeCrewMember, nudgeAllPending,
  sendNudgeEmail (currently RSVP-shaped; will re-add targeted at
  non-signed-up invitees when we rebuild).
- Dead helpers: sendRsvpBlastExistingUser, sendRsvpBlastNewUser.
- updateActionCenterSettings now narrows to travelEnabled only.
- Crew tab RSVP UI: chips, tallies, background tint, per-row override,
  bulk "Nudge pending" button. Crew tab now renders identically across
  planning and going stages.
- e2e/rsvp.spec.ts and the setRsvpStatus / setRsvpStatusForMember test
  cases; travel tests kept under a renamed describe block.

Added — InvitationCard (renamed from RsvpActionCard)
- Invitation message + Edit / Default-invite controls (as before).
- Recipients grid below the invitation: name + email per member,
  1→2→3→4 responsive cols, RoleBadge next to planners/owner, sorted
  the same way the Crew tab does.
- Inline ghost chips for crew missing an email — tapping a chip swaps
  the chip for a name + email input + Save/Cancel, persisted via
  ghostCrew.update.
- Send invitation button moved below the recipient list and shows the
  count of addressable recipients; disabled when zero.
- Travel form and the Share-travel-info toggle preserved.

Left untouched
- DB columns trip_members.rsvp_status and trips.rsvp_enabled — cheap
  to re-enable later, painful to re-add if we change our minds.
- The rsvp_response notification enum entry in notificationText.ts
  and TopNav.tsx — harmless, unreachable in prod now.
