+++ src/app/trips/[tripId]/tabs/CrewTab.tsx
@@ -1,27 +1,13 @@
 "use client";
 
 import { useState } from "react";
-import { Ghost, Send, X } from "lucide-react";
+import { Ghost, X } from "lucide-react";
 import { CrewSearchInput } from "@/components/CrewSearchInput";
 import { UserAvatar } from "@/components/UserAvatar";
 import { useTheme } from "next-themes";
 import { trpc } from "@/lib/trpc-client";
 import { useCurrentUser } from "@/hooks/useCurrentUser";
 import type { TabProps } from "./types";
-
-// ── RSVP helpers ──────────────────────────────────────────────────────────
-
-const RSVP_LABEL: Record<string, { label: string; color: string; bg: string }> = {
-  in:      { label: "In",    color: "var(--color-bt-accent)", bg: "var(--color-bt-accent-faint, rgba(0,212,170,0.1))" },
-  maybe:   { label: "Maybe", color: "var(--color-bt-warning)", bg: "var(--color-bt-warning-faint, rgba(245,158,11,0.1))" },
-  out:     { label: "Out",   color: "var(--color-bt-danger)", bg: "var(--color-bt-danger-faint, rgba(239,68,68,0.1))" },
-};
-
-const RSVP_SORT_ORDER: Record<string, number> = { in: 0, maybe: 1, out: 3 };
-function _rsvpSortKey(rsvpStatus: string | null): number {
-  if (rsvpStatus === null || rsvpStatus === undefined) return 2; // pending after maybe, before out
-  return RSVP_SORT_ORDER[rsvpStatus] ?? 2;
-}
 
 const ROLE_ORDER: Record<string, number> = { Owner: 0, Planner: 1, Member: 2 };
 
@@ -32,7 +18,6 @@ type Member = {
   user_id: string | null;
   role: string;
   status: string | null;
-  rsvp_status?: string | null;
   displayName: string;
   isGuest: boolean;
   user: { email: string | null; is_guest?: boolean } | null;
@@ -48,8 +33,6 @@ function CrewMemberRow({
   isMe,
   isExpanded,
   index,
-  showRsvpStatus,
-  stage,
   isPlannerRow,
   onToggle,
   onUpdated,
@@ -61,8 +44,6 @@ function CrewMemberRow({
   isMe: boolean;
   isExpanded: boolean;
   index: number;
-  showRsvpStatus: boolean;
-  stage: string;
   isPlannerRow?: boolean;
   onToggle: () => void;
   onUpdated: () => void;
@@ -87,22 +68,7 @@ function CrewMemberRow({
     onSuccess() { utils.tripMembers.list.invalidate({ tripId }); },
   });
 
-  const nudgeCrewMember = trpc.tripMembers.nudgeCrewMember.useMutation();
-  const setRsvpForMember = trpc.tripMembers.setRsvpStatusForMember.useMutation({
-    onSuccess() { utils.tripMembers.list.invalidate({ tripId }); },
-  });
-  const [nudgeSent, setNudgeSent] = useState(false);
-
-  const handleNudge = async (e: React.MouseEvent) => {
-    e.stopPropagation();
-    if (!m.user_id) return;
-    await nudgeCrewMember.mutateAsync({ tripId, userId: m.user_id });
-    setNudgeSent(true);
-    setTimeout(() => setNudgeSent(false), 2000);
-  };
-
   const display = m.displayName;
-  const rsvpCfg = m.rsvp_status ? RSVP_LABEL[m.rsvp_status] : null;
   const editable = canEdit && !isMe && m.role !== "Owner";
   const canActOnPlanner = isOwner && !isMe && m.role === "Planner";
   const canPromoteToCoplanner = isOwner && !isMe && m.role === "Member" && !isPlannerRow && !!m.user?.email;
@@ -146,7 +112,6 @@ function CrewMemberRow({
       <div
         className="flex items-center gap-3 py-2.5 px-1 -mx-1 rounded"
         style={{
-          background: rsvpCfg && !m.isGuest ? `${rsvpCfg.color}0a` : undefined,
           cursor: editable && !isPlannerRow ? "pointer" : undefined,
         }}
         onClick={editable && !isPlannerRow ? onToggle : undefined}
@@ -182,34 +147,8 @@ function CrewMemberRow({
           ) : null}
         </div>
 
-        {/* ── Right side: status + actions ─────────────────────────────────── */}
+        {/* ── Right side: actions ──────────────────────────────────────────── */}
         <div className="flex flex-shrink-0 items-center">
-          {/* RSVP chip — going stage only */}
-          {showRsvpStatus && (
-          <div className="flex w-[64px] flex-shrink-0 justify-center">
-            {(() => {
-              const rsvp = m.rsvp_status;
-              const cfg = rsvp ? RSVP_LABEL[rsvp] : null;
-              return cfg ? (
-                <span
-                  className="rounded-full px-2 py-0.5 text-xs"
-                  style={{ background: cfg.bg, color: cfg.color }}
-                >
-                  {cfg.label}
-                </span>
-              ) : (
-                <span
-                  className="rounded-full px-2 py-0.5 text-xs"
-                  style={{ background: "var(--color-bt-card-raised)", color: "var(--color-bt-text-dim)" }}
-                >
-                  Pending
-                </span>
-              );
-            })()}
-          </div>
-          )}
-
-          {/* Col 3: actions */}
           <div className="flex flex-shrink-0 items-center gap-1">
             {/* Planner row: move to rest of crew */}
             {isPlannerRow && canActOnPlanner && (
@@ -301,64 +240,6 @@ function CrewMemberRow({
             </div>
           )}
 
-          {/* RSVP override + nudge (going stage) */}
-          {showRsvpStatus && (
-            <div>
-              <p className="mb-1 text-xs" style={{ color: "var(--color-bt-text-dim)" }}>RSVP</p>
-              <div className="flex flex-wrap items-center gap-1.5">
-                {(["in", "maybe", "out"] as const).map((val) => {
-                  const cfg = RSVP_LABEL[val];
-                  const isSelected = m.rsvp_status === val;
-                  return (
-                    <button
-                      key={val}
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        if (m.user_id) setRsvpForMember.mutate({ tripId, userId: m.user_id, rsvpStatus: val });
-                      }}
-                      disabled={setRsvpForMember.isPending}
-                      className="rounded-lg px-2.5 py-1 text-xs font-medium transition-colors disabled:opacity-40"
-                      style={{
-                        background: isSelected ? cfg.bg : "var(--color-bt-base)",
-                        color: isSelected ? cfg.color : "var(--color-bt-text-dim)",
-                        border: `1px solid ${isSelected ? cfg.color : "var(--color-bt-border)"}`,
-                      }}
-                    >
-                      {cfg.label}
-                    </button>
-                  );
-                })}
-                {m.rsvp_status && (
-                  <button
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      if (m.user_id) setRsvpForMember.mutate({ tripId, userId: m.user_id, rsvpStatus: null });
-                    }}
-                    disabled={setRsvpForMember.isPending}
-                    className="rounded-lg px-2.5 py-1 text-xs disabled:opacity-40"
-                    style={{ color: "var(--color-bt-text-dim)", border: "1px solid var(--color-bt-border)" }}
-                  >
-                    Clear
-                  </button>
-                )}
-                {/* Nudge — only when member has email and hasn't confirmed */}
-                {m.user?.email && m.role !== "Owner" && (m.rsvp_status == null || m.rsvp_status === "maybe") && (
-                  <button
-                    onClick={handleNudge}
-                    disabled={nudgeCrewMember.isPending || nudgeSent}
-                    className="ml-auto flex items-center gap-1 rounded-lg px-2.5 py-1 text-xs disabled:opacity-40"
-                    style={{
-                      color: nudgeSent ? "var(--color-bt-accent)" : "var(--color-bt-text-dim)",
-                      border: "1px solid var(--color-bt-border)",
-                    }}
-                  >
-                    {nudgeSent ? "Sent ✓" : <><Send size={11} /><span>Nudge</span></>}
-                  </button>
-                )}
-              </div>
-            </div>
-          )}
-
           {/* Actions row — save/cancel (planner promotion lives in the inline row button) */}
           <div className="flex items-center justify-end gap-2">
             {m.isGuest && (
@@ -400,8 +281,6 @@ export function CrewTab({ trip, canEdit }: TabProps) {
 
   const createGhost = trpc.ghostCrew.create.useMutation();
   const inviteByEmail = trpc.tripMembers.inviteByEmail.useMutation();
-  const nudgeAll = trpc.tripMembers.nudgeAllPending.useMutation();
-  const [nudgeAllSent, setNudgeAllSent] = useState(false);
 
   const me = members.find((m) => m.user_id === currentUser?.id);
   const isOwner = me?.role === "Owner";
@@ -436,39 +315,12 @@ export function CrewTab({ trip, canEdit }: TabProps) {
     }
   };
 
-  const stage = trip.stage ?? "idea";
-  const showRsvpStatus = stage === "going";
-
-  // Count pending members who have an email (can actually be nudged)
-  const pendingWithEmailCount = members.filter(
-    (m) => (m as { rsvp_status?: string | null }).rsvp_status == null && !!m.user?.email
-  ).length;
-
   // Split members into planners and crew
   const plannersSorted = sorted.filter((m) => m.role === "Owner" || m.role === "Planner");
   const crewSorted = sorted.filter((m) => m.role !== "Owner" && m.role !== "Planner");
 
   return (
     <div className="space-y-4 px-4">
-      {/* Nudge button (going stage only) */}
-      {canEdit && showRsvpStatus && pendingWithEmailCount > 0 && (
-        <div className="flex items-center justify-center">
-          <button
-            onClick={async () => {
-              await nudgeAll.mutateAsync({ tripId });
-              setNudgeAllSent(true);
-              setTimeout(() => setNudgeAllSent(false), 3000);
-            }}
-            disabled={nudgeAll.isPending || nudgeAllSent}
-            className="flex items-center gap-1.5 rounded-lg px-3 py-1.5 text-xs font-medium disabled:opacity-40"
-            style={{ border: "1px solid var(--color-bt-border)", color: "var(--color-bt-text-dim)" }}
-          >
-            <Send size={13} />
-            {nudgeAllSent ? "Sent ✓" : `Nudge pending (${pendingWithEmailCount})`}
-          </button>
-        </div>
-      )}
-
       {/* ── PLANNERS section ── */}
       <div>
         <h2
@@ -509,8 +361,6 @@ export function CrewTab({ trip, canEdit }: TabProps) {
               isMe={isMe}
               index={i}
               isExpanded={false}
-              showRsvpStatus={showRsvpStatus}
-              stage={stage}
               isPlannerRow
               onToggle={() => {}}
               onUpdated={() => utils.tripMembers.list.invalidate({ tripId })}
@@ -564,22 +414,6 @@ export function CrewTab({ trip, canEdit }: TabProps) {
           </div>
         )}
 
-        {/* RSVP headcount summary (GOING/NOW stage) */}
-        {showRsvpStatus && (() => {
-          const inC = members.filter((m) => (m as { rsvp_status?: string | null }).rsvp_status === "in").length;
-          const maybeC = members.filter((m) => (m as { rsvp_status?: string | null }).rsvp_status === "maybe").length;
-          const outC = members.filter((m) => (m as { rsvp_status?: string | null }).rsvp_status === "out").length;
-          const pendC = members.filter((m) => (m as { rsvp_status?: string | null }).rsvp_status == null).length;
-          return (
-            <p
-              className="mb-3 inline-block rounded-full px-3 py-1.5 text-[13px]"
-              style={{ background: "var(--color-bt-card-raised)", color: "var(--color-bt-text-dim)" }}
-            >
-              {inC} in · {maybeC} maybe · {pendC} pending · {outC} out
-            </p>
-          );
-        })()}
-
         <div>
           {crewSorted.map((m, i) => {
             const isMe = m.user_id === currentUser?.id;
@@ -593,8 +427,6 @@ export function CrewTab({ trip, canEdit }: TabProps) {
                 isMe={isMe}
                 index={i}
                 isExpanded={expandedId === m.user_id}
-                showRsvpStatus={showRsvpStatus}
-                stage={stage}
                 onToggle={() => setExpandedId(expandedId === m.user_id ? null : m.user_id)}
                 onUpdated={() => utils.tripMembers.list.invalidate({ tripId })}
               />
