+++ src/server/routers/tripMembers.ts
@@ -16,7 +16,7 @@ export const tripMembersRouter = router({
     .query(async ({ ctx }) => {
       const { data, error } = await ctx.supabase
         .from("trip_members")
-        .select("id, trip_id, user_id, role, status, rsvp_status, joined_at, travel_mode, travel_detail, flight_airline, flight_number, flight_arrival_time, flight_airport, travel_shared")
+        .select("id, trip_id, user_id, role, status, joined_at, travel_mode, travel_detail, flight_airline, flight_number, flight_arrival_time, flight_airport, travel_shared")
         .eq("trip_id", ctx.tripId)
         .order("joined_at", { ascending: true });
 
@@ -419,148 +419,6 @@ export const tripMembersRouter = router({
     }),
 
   // -----------------------------------------------------------------------
-  // updateRsvp — any real member can update their own RSVP status
-  // Ghost crew always stay "in" — no RSVP for ghosts.
-  // -----------------------------------------------------------------------
-  updateRsvp: authedProcedure
-    .input(
-      z.object({
-        tripId: z.string(),
-        status: z.enum(["in", "likely", "maybe", "out"]),
-      })
-    )
-    .use(requireTripMember)
-    .mutation(async ({ ctx, input }) => {
-      const { data, error } = await ctx.supabase
-        .from("trip_members")
-        .update({ status: input.status })
-        .eq("trip_id", ctx.tripId)
-        .eq("user_id", ctx.user!.id)
-        .select()
-        .single();
-
-      if (error) {
-        throw new TRPCError({
-          code: "INTERNAL_SERVER_ERROR",
-          message: "Failed to update RSVP",
-        });
-      }
-
-      return data;
-    }),
-
-  // -----------------------------------------------------------------------
-  // setRsvpStatus — member sets their own RSVP status
-  // Only works when trip is in 'going' stage
-  // -----------------------------------------------------------------------
-  setRsvpStatus: authedProcedure
-    .input(
-      z.object({
-        tripId: z.string(),
-        rsvpStatus: z.enum(["in", "maybe", "out"]),
-      })
-    )
-    .use(requireTripMember)
-    .mutation(async ({ ctx, input }) => {
-      // Verify trip is in going stage
-      const { data: trip } = await ctx.supabase
-        .from("trips")
-        .select("stage, title")
-        .eq("id", ctx.tripId)
-        .single();
-
-      if (!trip || trip.stage !== "going") {
-        throw new TRPCError({
-          code: "BAD_REQUEST",
-          message: "RSVP can only be set when the trip is in the going stage.",
-        });
-      }
-
-      const { data, error } = await ctx.supabase
-        .from("trip_members")
-        .update({ rsvp_status: input.rsvpStatus })
-        .eq("trip_id", ctx.tripId)
-        .eq("user_id", ctx.user!.id)
-        .select("user_id, trip_id, rsvp_status")
-        .single();
-
-      if (error || !data) {
-        throw new TRPCError({
-          code: "INTERNAL_SERVER_ERROR",
-          message: "Failed to update RSVP status",
-        });
-      }
-
-      // Fire notification to owner (unless owner is setting their own RSVP)
-      try {
-        const { data: ownerMember } = await ctx.supabase
-          .from("trip_members")
-          .select("user_id")
-          .eq("trip_id", ctx.tripId)
-          .eq("role", "Owner")
-          .single();
-
-        if (ownerMember && ownerMember.user_id !== ctx.user!.id) {
-          const { data: currentUserData } = await ctx.supabase
-            .from("users")
-            .select("name")
-            .eq("id", ctx.user!.id)
-            .single();
-
-          const { createNotification } = await import("./notifications");
-          await createNotification(ctx.supabase, {
-            tripId: ctx.tripId,
-            actorId: ctx.user!.id,
-            recipientId: ownerMember.user_id,
-            type: "rsvp_response",
-            payload: {
-              responder_name: currentUserData?.name ?? "Someone",
-              rsvp_status: input.rsvpStatus,
-              trip_name: trip.title,
-              trip_id: ctx.tripId,
-            },
-          });
-        }
-      } catch {
-        // Notification failure shouldn't block the mutation
-      }
-
-      return data;
-    }),
-
-  // -----------------------------------------------------------------------
-  // setRsvpStatusForMember — owner/planner sets RSVP for any member
-  // Used for ghost members who have no app access
-  // -----------------------------------------------------------------------
-  setRsvpStatusForMember: authedProcedure
-    .input(
-      z.object({
-        tripId: z.string(),
-        userId: z.string(),
-        rsvpStatus: z.enum(["in", "maybe", "out"]).nullable(),
-      })
-    )
-    .use(requireTripRole("Planner"))
-    .mutation(async ({ ctx, input }) => {
-      const { data, error } = await ctx.supabase
-        .from("trip_members")
-        .update({ rsvp_status: input.rsvpStatus })
-        .eq("trip_id", ctx.tripId)
-        .eq("user_id", input.userId)
-        .select("user_id, trip_id, rsvp_status")
-        .single();
-
-      if (error || !data) {
-        throw new TRPCError({
-          code: "INTERNAL_SERVER_ERROR",
-          message: "Failed to update RSVP status",
-        });
-      }
-
-      return data;
-    }),
-
-  // -----------------------------------------------------------------------
   // updateTravel — member updates their own travel info
   // -----------------------------------------------------------------------
   updateTravel: authedProcedure
@@ -604,140 +462,6 @@ export const tripMembersRouter = router({
       }
 
       return data;
-    }),
-
-  // -----------------------------------------------------------------------
-  // nudgeCrewMember — resend RSVP email to a specific crew member
-  // -----------------------------------------------------------------------
-  nudgeCrewMember: authedProcedure
-    .input(
-      z.object({
-        tripId: z.string(),
-        userId: z.string(),
-      })
-    )
-    .use(requireTripRole("Planner"))
-    .mutation(async ({ ctx, input }) => {
-      // Verify trip is in going stage
-      const { data: trip } = await ctx.supabase
-        .from("trips")
-        .select("stage, title")
-        .eq("id", ctx.tripId)
-        .single();
-
-      if (!trip || trip.stage !== "going") {
-        throw new TRPCError({ code: "BAD_REQUEST", message: "Trip is not in going stage." });
-      }
-
-      // Fetch target member + user data
-      const { data: member } = await ctx.supabase
-        .from("trip_members")
-        .select("user_id, rsvp_status, users(name, email, is_guest)")
-        .eq("trip_id", ctx.tripId)
-        .eq("user_id", input.userId)
-        .single();
-
-      if (!member) {
-        throw new TRPCError({ code: "NOT_FOUND", message: "Member not found." });
-      }
-
-      const user = member.users as unknown as { name: string | null; email: string | null; is_guest: boolean } | null;
-
-      if (!user?.email) {
-        throw new TRPCError({ code: "BAD_REQUEST", message: "Member has no email on file." });
-      }
-
-      if (member.rsvp_status === "in" || member.rsvp_status === "out") {
-        throw new TRPCError({ code: "BAD_REQUEST", message: "Member has already responded." });
-      }
-
-      // Fetch inviter name
-      const { data: inviter } = await ctx.supabase
-        .from("users")
-        .select("name, nickname")
-        .eq("id", ctx.user!.id)
-        .single();
-
-      try {
-        const { sendNudgeEmail } = await import("@/lib/email");
-        await sendNudgeEmail({
-          toEmail: user.email,
-          toName: user.name ?? user.email.split("@")[0],
-          inviterName: inviter?.nickname ?? inviter?.name ?? "Your trip organizer",
-          tripName: trip.title,
-          tripId: ctx.tripId,
-        });
-      } catch {
-        throw new TRPCError({ code: "INTERNAL_SERVER_ERROR", message: "Failed to send nudge email." });
-      }
-
-      return { success: true };
-    }),
-
-  // -----------------------------------------------------------------------
-  // nudgeAllPending — send nudge to all pending members with email
-  // -----------------------------------------------------------------------
-  nudgeAllPending: authedProcedure
-    .input(z.object({ tripId: z.string() }))
-    .use(requireTripRole("Planner"))
-    .mutation(async ({ ctx }) => {
-      // Verify trip is in going stage
-      const { data: trip } = await ctx.supabase
-        .from("trips")
-        .select("stage, title")
-        .eq("id", ctx.tripId)
-        .single();
-
-      if (!trip || trip.stage !== "going") {
-        throw new TRPCError({ code: "BAD_REQUEST", message: "Trip is not in going stage." });
-      }
-
-      // Fetch all pending members (rsvp_status IS NULL) with user details
-      const { data: members } = await ctx.supabase
-        .from("trip_members")
-        .select("user_id, rsvp_status, users(name, email, is_guest)")
-        .eq("trip_id", ctx.tripId)
-        .is("rsvp_status", null);
-
-      if (!members || members.length === 0) {
-        return { sent: 0, skipped: 0 };
-      }
-
-      // Fetch inviter name
-      const { data: inviter } = await ctx.supabase
-        .from("users")
-        .select("name, nickname")
-        .eq("id", ctx.user!.id)
-        .single();
-
-      const inviterName = inviter?.nickname ?? inviter?.name ?? "Your trip organizer";
-      let sent = 0;
-      let skipped = 0;
-
-      const { sendNudgeEmail } = await import("@/lib/email");
-
-      for (const member of members) {
-        const user = member.users as unknown as { name: string | null; email: string | null; is_guest: boolean } | null;
-        if (!user?.email) {
-          skipped++;
-          continue;
-        }
-
-        try {
-          await sendNudgeEmail({
-            toEmail: user.email,
-            toName: user.name ?? user.email.split("@")[0],
-            inviterName,
-            tripName: trip.title,
-            tripId: ctx.tripId,
-          });
-          sent++;
-        } catch {
-          skipped++;
-        }
-      }
-
-      return { sent, skipped };
     }),
 
   // -----------------------------------------------------------------------
